RFC: Dynamic Provider Instances and Instance Assignment
This RFC describes an alternative approach for issue #300 that treats it
as a dynamic evaluation problem rather than as an early-eval problem.

For the moment my goal of this proposal is not to actually replace the
existing early eval proposal, but rather to draw attention to some aspects
of it that would be difficult to translate to fully-dynamic eval if we
heard requests for that later.

This document therefore aims to ask the question of firstly whether we'd
like to impose some similar constraints on the early eval approach even
though they aren't strictly needed there, and secondly whether this
realization that we don't have any current experience retroactively
changing a feature from early eval to dynamic eval represents enough risk
that we might prefer to put even the early-eval approach on pause while
we analyze the potential risks in more detail.

It's likely that if we _did_ choose to implement the dynamic approach we'd
want to refine this RFC further to draw out some more detail in the
implementation section. I've intentionally skimped on that section here
because the priority is to have the conversation about the potential new
user-facing constraints, and not to actually implement what this proposal
describes in the near future. The implementation details are included as
a starting point, but are insufficiently detailed to actually begin
implementation.